Fixed stuff

+++ src/video/color.hpp
@@ -84,8 +84,8 @@ public:
     return greyscale() < other.greyscale();
   }
 
-  std::vector<int> toVector(bool include_alpha){
-    std::vector<int> result;
+  std::vector<float> toVector(bool include_alpha){
+    std::vector<float> result;
     result.clear();
     result.push_back(red);
     result.push_back(green);
